regulatory: fix world regulatory domain data
The precise rule definition here isn't really valid, it would
be rejected if it came from userspace due to the bandwidth it
specifies being bigger than the rule's width.

This is fairly much inconsequential since the other rule before
it does enable the bandwidth, but express that better by using
the NL80211_RRF_AUTO_BW flag.

Signed-off-by: Johannes Berg <johannes.berg@intel.com>