Remove EFI memmap quirk for UV2+
Commit a5d90c923bcf ("x86/efi: Quirk out SGI UV") added a quirk to
efi_apply_memmap_quirks to force SGI UV systems to fall back to the old
EFI memmap mechanism.  We have a BIOS fix for this issue on all systems
except for UV1.  This commit fixes up the EFI quirk/MMR mapping code so
that we only apply the special case to UV1 hardware.

Signed-off-by: Alex Thorlton <athorlton@sgi.com>
Cc: Thomas Gleixner <tglx@linutronix.de>
Cc: Ingo Molnar <mingo@redhat.com>
Cc: "H. Peter Anvin" <hpa@zytor.com>
Cc: x86@kernel.org
Cc: Matt Fleming <matt@codeblueprint.co.uk>
Cc: Dimitri Sivanich <sivanich@sgi.com>
Cc: Hedi Berriche <hedi@sgi.com>
Cc: Mike Travis <travis@sgi.com>
Cc: Len Brown <len.brown@intel.com>
Cc: linux-efi@vger.kernel.org