ntp:Change the type of time_reftime to time64_t and replace get_seconds with __ktime_get_real_seconds
The type of static variant "time_reftime" and the call of get_seconds
in ntp are both not y2038 safe.
So change the type of time_reftime to time64_t and replace get_seconds
with __ktime_get_real_seconds.
The local variant "secs" in ntp_update_offset represents seconds between
now and last ntp adjustment, it seems impossible that this time will last
more than 68 years, so keep its type as "long".

Reviewed-by: John Stultz <john.stultz@linaro.org>
Signed-off-by: DengChao <chao.deng@linaro.org>