block: make genhd.c slightly more explicitly non-modular
The Makefile currently controlling compilation of this code is:

obj-$(CONFIG_BLOCK) := bio.o elevator.o blk-core.o blk-tag.o blk-sysfs.o \
                        [...]
                        genhd.o scsi_ioctl.o partition-generic.o ioprio.o \

...and the Kconfig is:

menuconfig BLOCK
       bool "Enable the block layer" if EXPERT
       default y

...meaning that it currently is not being built as a module by anyone.

While the code uses moduleparam and other modular infrastructure that
prevents removal of module.h itself, we can at least ensure the way
it is loaded sends a clear message that it is built-in or absent.

Since module_init translates to device_initcall in the non-modular
case, the init ordering remains unchanged with this commit.  We
do that because the file already has subsys_initcall() usage within,
and I don't want to risk breaking ordering within the file itself.

Cc: Jens Axboe <axboe@kernel.dk>
Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>