regmap: flat: introduce register striding to save some memories
Throughout the regcache-flat code, for the flat cache array, the
actual register address offsets are used to index the values, and
this will waste some memory spaces.

For example, on 64-BIT platform, device A has three registers:
    register offsets: {0x00, 0x08, 0x10}
    max_register: 0x10
    regsiter striding: 8

And the size of flat cache memory space will be:
    (max_register + 1 ) * sizeof(unsigned int) =
    (0x10 + 1) * sizeof(unsigned int) = 17 * 8 = 136 Bytes

Since map->reg_stride has been introduced and all extant register
addresses are a multiple of this value, it could use the address
offsets divide by the stride to determine the index.

Then the size of flat cache memory space will be:
    (max_register / reg_stride + 1 ) * sizeof(unsigned int) =
    (0x10 / 8 + 1) * sizeof(unsigned int) = 3 * 8 = 24 Bytes

And the bigger of the striding value, there will be more memory
space wasted. After introducing the register striding here can
save some memeories for the system.

Signed-off-by: Xiubo Li <lixiubo@cmss.chinamobile.com>