6lowpan: iphc: add support for stateful compression
This patch introduce support for IPHC stateful address compression. It
will offer the context table via one debugfs entry.

Example to setup a context id:

A "cat /sys/kernel/debug/6lowpan/lowpan0/ctx_table" will display all
contexts which are available. Example:

ID ipv6-address/prefix-length                  flags
0  0000:0000:0000:0000:0000:0000:0000:0000/0   0
1  0000:0000:0000:0000:0000:0000:0000:0000/0   0
2  0000:0000:0000:0000:0000:0000:0000:0000/0   0
3  0000:0000:0000:0000:0000:0000:0000:0000/0   0
4  0000:0000:0000:0000:0000:0000:0000:0000/0   0
5  0000:0000:0000:0000:0000:0000:0000:0000/0   0
6  0000:0000:0000:0000:0000:0000:0000:0000/0   0
7  0000:0000:0000:0000:0000:0000:0000:0000/0   0
8  0000:0000:0000:0000:0000:0000:0000:0000/0   0
9  0000:0000:0000:0000:0000:0000:0000:0000/0   0
10 0000:0000:0000:0000:0000:0000:0000:0000/0   0
11 0000:0000:0000:0000:0000:0000:0000:0000/0   0
12 0000:0000:0000:0000:0000:0000:0000:0000/0   0
13 0000:0000:0000:0000:0000:0000:0000:0000/0   0
14 0000:0000:0000:0000:0000:0000:0000:0000/0   0
15 0000:0000:0000:0000:0000:0000:0000:0000/0   0

For setting a context e.g. context id 0, context 2001::, prefix-length
64.

Hint: Simple copy one line and then maniuplate it.

echo "0 2001:0000:0000:0000:0000:0000:0000:0000/64 3" >
/sys/kernel/debug/6lowpan/lowpan0/ctx_table

The flags are currently two:

 - BIT(0) - active: entry is added or deleted to the ctx_table.
 - BIT(1) - c: compression flag according rfc6775.

On transmit side:

The IPHC code will automatically search for a context which would be
match for the address. Then it will be use the context with the
best compression method. Means the longest prefix which match will be
used.

Example:

2001::/126 vs 2001::/127 - the 2001::/127 can be full compressed if the
last bit of the address which has the prefix 2001::/127 is the same like
the IID from the Encapsulating Header. A context ID can also be a
2001::1/128, which is then a full ipv6 address.

On Receive side:

If there is a context defined (when CID not available then it's the
default context 0) then it will be used, if the header doesn't set
SAC or DAC bit thens, it will be dropped.

Signed-off-by: Alexander Aring <alex.aring@gmail.com>